refactor: migrate from Double to Float for SIMD/NEON optimization
- Replace Double with Float throughout audio engine (C++ and Kotlin)
- Update FrequencyPoint, FrequencyCurve, FrequencyRange to use Float
- Refactor GeneratorState to use Float phases for NEON compatibility
- Update lookup tables to use Float vectors
- Maintain sufficient precision for audio synthesis (Float is adequate for frequencies < 1000 Hz)

+++ core/audio/src/main/cpp/src/Wavetable.cpp
@@ -6,13 +6,13 @@ namespace binaural {
 std::vector<float> Wavetable::s_sineTable;
 int Wavetable::s_tableSize = DEFAULT_TABLE_SIZE;
 int Wavetable::s_tableSizeMask = DEFAULT_TABLE_SIZE - 1;
-double Wavetable::s_scaleFactor = DEFAULT_TABLE_SIZE / TWO_PI;
+float Wavetable::s_scaleFactor = DEFAULT_TABLE_SIZE / TWO_PI;
 float Wavetable::s_scaleFactorFloat = static_cast<float>(DEFAULT_TABLE_SIZE / TWO_PI);
 
 void Wavetable::initialize(int size) {
     s_tableSize = size;
     s_tableSizeMask = size - 1;
-    s_scaleFactor = static_cast<double>(size) / TWO_PI;
+    s_scaleFactor = static_cast<float>(size) / TWO_PI;
     s_scaleFactorFloat = static_cast<float>(s_scaleFactor);
     
     s_sineTable.resize(size);
